+++ packages/broker/test/integration/plugins/metrics/node/NodeMetrics.test.ts
@@ -32,7 +32,7 @@ describe('NodeMetrics', () => {
         streamIdPrefix = stream.id.replace('sec', '')
 
         metricsGeneratingBroker = await startBroker({
-            name: 'broker1',
+            name: 'metricsGeneratingBroker',
             privateKey: tmpAccount.privateKey,
             trackerPort,
             extraPlugins: {