+++ packages/broker/test/integration/plugins/storage/StorageConfig.test.ts
@@ -48,8 +48,8 @@ describe('StorageConfig', () => {
         })
     })
 
-    afterAll(() => {
-        cassandraClient.shutdown()
+    afterAll(async () => {
+        await cassandraClient?.shutdown()
     })
 
     beforeEach(async () => {